Continued move types -> core
