+++ core/src/server/resource_limiting.rs
@@ -52,7 +52,7 @@
 //! `#[method]` attribute:
 //!
 //! ```
-//! # use jsonrpsee::{types::RpcResult, proc_macros::rpc};
+//! # use jsonrpsee::{core::RpcResult, proc_macros::rpc};
 //! #
 //! #[rpc(server)]
 //! pub trait Rpc {
@@ -91,8 +91,8 @@
 
 use std::sync::Arc;
 
+use crate::Error;
 use arrayvec::ArrayVec;
-use jsonrpsee_types::error::Error;
 use parking_lot::Mutex;
 
 // The number of kinds of resources that can be used for limiting.
