+++ examples/proc_macro.rs
@@ -26,9 +26,8 @@
 
 use std::net::SocketAddr;
 
-use jsonrpsee::core::{async_trait, client::Subscription};
+use jsonrpsee::core::{async_trait, client::Subscription, Error};
 use jsonrpsee::proc_macros::rpc;
-use jsonrpsee::types::error::Error;
 use jsonrpsee::ws_client::WsClientBuilder;
 use jsonrpsee::ws_server::{SubscriptionSink, WsServerBuilder, WsServerHandle};
