+++ examples/ws_subscription.rs
@@ -28,8 +28,8 @@ use std::net::SocketAddr;
 
 use jsonrpsee::core::client::Subscription;
 use jsonrpsee::core::traits::SubscriptionClient;
+use jsonrpsee::core::Error;
 use jsonrpsee::rpc_params;
-use jsonrpsee::types::Error;
 use jsonrpsee::ws_client::WsClientBuilder;
 use jsonrpsee::ws_server::{RpcModule, WsServerBuilder};
