+++ http-client/src/client.rs
@@ -28,10 +28,11 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use crate::transport::HttpTransportClient;
-use crate::types::{Error, ErrorResponse, Id, NotificationSer, ParamsSer, RequestSer, Response, TEN_MB_SIZE_BYTES};
+use crate::types::{ErrorResponse, Id, NotificationSer, ParamsSer, RequestSer, Response, TEN_MB_SIZE_BYTES};
 use async_trait::async_trait;
 use jsonrpsee_core::client::{CertificateStore, RequestIdManager, Subscription};
 use jsonrpsee_core::traits::{Client, SubscriptionClient};
+use jsonrpsee_core::Error;
 use rustc_hash::FxHashMap;
 use serde::de::DeserializeOwned;
