+++ http-server/src/server.rs
@@ -36,13 +36,13 @@ use futures_util::{future::join_all, stream::StreamExt, FutureExt};
 use hyper::server::{conn::AddrIncoming, Builder as HyperBuilder};
 use hyper::service::{make_service_fn, service_fn};
 use hyper::Error as HyperError;
+use jsonrpsee_core::error::{Error, GenericTransportError};
 use jsonrpsee_core::http_helpers::read_body;
 use jsonrpsee_core::middleware::Middleware;
 use jsonrpsee_core::server::helpers::{collect_batch_response, prepare_error, MethodSink};
 use jsonrpsee_core::server::resource_limiting::Resources;
 use jsonrpsee_core::server::rpc_module::{MethodResult, Methods};
-use jsonrpsee_types::error::{Error, GenericTransportError};
-use jsonrpsee_types::error_response::ErrorCode;
+use jsonrpsee_types::error::ErrorCode;
 use jsonrpsee_types::{Id, Notification, Request, TEN_MB_SIZE_BYTES};
 use serde_json::value::RawValue;
 use socket2::{Domain, Socket, Type};
