+++ tests/tests/integration_tests.rs
@@ -32,11 +32,11 @@ use std::time::Duration;
 
 use helpers::{http_server, websocket_server, websocket_server_with_subscription};
 use jsonrpsee::core::client::Subscription;
+use jsonrpsee::core::error::SubscriptionClosedError;
 use jsonrpsee::core::traits::{Client, SubscriptionClient};
+use jsonrpsee::core::{Error, JsonValue};
 use jsonrpsee::http_client::HttpClientBuilder;
 use jsonrpsee::rpc_params;
-use jsonrpsee::types::error::SubscriptionClosedError;
-use jsonrpsee::types::{Error, JsonValue};
 use jsonrpsee::ws_client::WsClientBuilder;
 
 mod helpers;
