+++ tests/tests/resource_limiting.rs
@@ -28,10 +28,10 @@ use std::net::SocketAddr;
 use std::time::Duration;
 
 use jsonrpsee::core::traits::Client;
+use jsonrpsee::core::Error;
 use jsonrpsee::http_client::HttpClientBuilder;
 use jsonrpsee::http_server::{HttpServerBuilder, HttpServerHandle};
 use jsonrpsee::proc_macros::rpc;
-use jsonrpsee::types::Error;
 use jsonrpsee::ws_client::WsClientBuilder;
 use jsonrpsee::ws_server::{WsServerBuilder, WsServerHandle};
 use jsonrpsee::RpcModule;
