+++ types/src/lib.rs
@@ -34,9 +34,6 @@ extern crate alloc;
 /// Ten megabytes.
 pub const TEN_MB_SIZE_BYTES: u32 = 10 * 1024 * 1024;
 
-/// Error type.
-pub mod error;
-
 /// JSON-RPC params related types.
 pub mod params;
 
@@ -47,22 +44,11 @@ pub mod request;
 pub mod response;
 
 /// JSON-RPC response error object related types.
-pub mod error_response;
+pub mod error;
 
 // pub use async_trait::async_trait;
-pub use beef::Cow;
-pub use error::{CallError, Error};
+pub use error::ErrorResponse;
 pub use params::{Id, Params, ParamsSequence, ParamsSer, SubscriptionId, TwoPointZero};
-pub use serde::{de::DeserializeOwned, Serialize};
-pub use serde_json::{
-	to_value as to_json_value, value::to_raw_value as to_json_raw_value, value::RawValue as JsonRawValue,
-	Value as JsonValue,
-};
-
-/// JSON-RPC result.
-pub type RpcResult<T> = std::result::Result<T, Error>;
-
-pub use error_response::ErrorResponse;
 pub use request::{InvalidRequest, Notification, NotificationSer, Request, RequestSer};
 pub use response::{Response, SubscriptionPayload, SubscriptionResponse};
