+++ ws-server/src/tests.rs
@@ -29,17 +29,16 @@ use std::fmt;
 use std::net::SocketAddr;
 use std::time::Duration;
 
-use crate::types::error::{CallError, Error};
-use crate::types::DeserializeOwned;
+use crate::types::error::CallError;
 use crate::types::{self, ErrorResponse, Response};
 use crate::{future::ServerHandle, RpcModule, WsServerBuilder};
 use anyhow::anyhow;
 use futures_util::future::join;
+use jsonrpsee_core::{to_json_raw_value, DeserializeOwned, Error};
 use jsonrpsee_test_utils::helpers::*;
 use jsonrpsee_test_utils::mocks::{Id, TestContext, WebSocketTestClient, WebSocketTestError};
 use jsonrpsee_test_utils::TimeoutFutureExt;
-use jsonrpsee_types::error_response::invalid_subscription_err;
-use jsonrpsee_types::to_json_raw_value;
+use jsonrpsee_types::error::invalid_subscription_err;
 use serde_json::Value as JsonValue;
 use tracing_subscriber::{EnvFilter, FmtSubscriber};
